requirements.txt
Pillow>=9.0.0
requests[socks]>=2.28.0
beautifulsoup4>=4.10.0